requirements.txt
asgiref==3.5.2
Django==4.1
django-crispy-forms==1.14.0
sqlparse==0.4.2
django-filter